use long for history archive indexes

+++ src/cpp/r/include/r/session/RConsoleHistory.hpp
@@ -47,7 +47,7 @@ private:
    // COPYING: boost::noncopyable
       
 public:   
-   void setCapacity(int capacity);
+   void setCapacity(long capacity);
 
    void setRemoveDuplicates(bool removeDuplicates);
    
@@ -56,15 +56,15 @@ public:
    const_iterator begin() const { return historyBuffer_.begin(); }
    const_iterator end() const { return historyBuffer_.end(); }
    
-   int size() const
+   long size() const
    {
       return historyBuffer_.size();
    }
 
    void clear();
 
-   void subset(int beginIndex, // inclusive
-               int endIndex,   // exclusive,
+   void subset(long beginIndex, // inclusive
+               long endIndex,   // exclusive,
                std::vector<std::string>* pEntries) const;
 
    void asJson(core::json::Array* pHistoryArray) const;
